plugins/classifier/requirements/classifier.cpu.requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.29
python-multipart==0.0.9
huggingface-hub==0.30.0
scipy==1.13.0
opencv-python==4.9.0.80
hdbscan==0.8.40
umap-learn==0.5.7
tqdm==4.67.1
matplotlib==3.10.0
PyYAML==6.0.2
timm==1.0.12
aiofiles==24.1.0